SISVGABlankScreen: only update the hardware if we're changing the state
The command `xset s reset` will reset the X server screensaver to
reset the timer for the screensaver and dpms.  If the display is
already on, it should do nothing as far as the hardware goes.  In the
SISVGABlankScreen routine it requires as reset to update the register,
but that will disrupt the VGA signal and cause the monitor to resync
leaving the screen blank for a couple seconds.  I have a program that
wants to keep the screensaver from coming on, but allow normal
operation when it isn't running.

Debian bug#546838 <http://bugs.debian.org/546838>

Signed-off-by: Julien Cristau <jcristau@debian.org>